sua 9 10

+++ Day 2/Bai9/src/bai9/Bai9.java
@@ -3,8 +3,8 @@ import java.util.*;
 public class Bai9 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        double a = sc.nextInt();
+        double b = sc.nextInt();
         if (a == 0){
             if (b == 0) System.out.println("Vo so nghiem");
             else System.out.println("Khong co nghiem");
